fix: fix some bugs

+++ services/workshop.service.js
@@ -81,6 +81,10 @@ class WorkshopService {
   async addUserToWorkshop(workshopId, userId) {
     try {
 
+      const user = await WorkshopDAO.getWorkshopUsers(workshopId, userId);
+      if (user) {
+        return { message: "User already exists in the workshop" };
+      }
       const workshopUsers = await WorkshopDAO.addUserToWorkshop(workshopId, userId);
       
       return WorkshopUsersDTO.fromDatabase(workshopUsers);
@@ -90,13 +94,29 @@ class WorkshopService {
     }
   }
 
-  async removeUserFromWorkshop(workshopId, userId) {
+  async removeUserFromWorkshop(workshopId, userId, currentUserId) {
     try {
-      const workshopUsers = await WorkshopDAO.removeUserFromWorkshop(workshopId, userId);
-      if (!workshopUsers) {
-        return { status: 404, message: "User not found in workshop" };
+      // First check if the user exists in the workshop
+      const workshopUser = await WorkshopDAO.getWorkshopUsers(workshopId, userId);
+      
+      if (!workshopUser) {
+        throw new Error("User is not a member of this workshop", 404);
       }
-      return WorkshopUsersDTO.fromDatabase(workshopUsers);
+
+      // Check if current user is admin or the same user
+      const currentUserWorkshop = await WorkshopDAO.getWorkshopUsers(workshopId, currentUserId);
+      
+      if (!currentUserWorkshop) {
+        throw new Error("You are not a member of this workshop", 403);
+      }
+
+      if (!currentUserWorkshop.role == 'ADMIN' && currentUserId !== userId) {
+        throw new Error("You don't have permission to remove other users", 403);
+      }
+
+      // Proceed with removal
+      const removedUser = await WorkshopDAO.removeUserFromWorkshop(workshopId, userId);
+      return WorkshopUsersDTO.fromDatabase(removedUser);
     } catch (error) {
       console.error("Error removing user from workshop:", error);
       throw error;
